Delete deadline with wolverine; Small refactorings
