+++ Misa.Application/Items/Commands/UpdateItemHandler.cs
@@ -1,4 +1,4 @@
-﻿using Misa.Application.Items.Repositories;
+﻿using Misa.Application.Common.Abstractions.Persistence;
 using Misa.Contract.Items;
 
 namespace Misa.Application.Items.Commands;
