+++ Misa.Application/Scheduling/Commands/Deadlines/RemoveItemDeadlineCommand.cs
@@ -1,0 +1,3 @@
+﻿namespace Misa.Application.Scheduling.Commands.Deadlines;
+
+public record RemoveItemDeadlineCommand(Guid ItemId);
