--- Misa.Application/Scheduling/Commands/UpsertItemDeadline/RemoveItemDeadlineCommand.cs
@@ -1,3 +1,0 @@
-﻿namespace Misa.Application.Scheduling.Commands.UpsertItemDeadline;
-
-public record RemoveItemDeadlineCommand(Guid ItemId);
